% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{clean_text}
\alias{clean_text}
\title{Clean/standardize text}
\usage{
clean_text(string)
}
\arguments{
\item{string}{a character string to "clean"}
}
\value{
the cleaned character string (no punctuation, spaces or capital letters)
}
\description{
Removes punctuation and spaces from character string. Also makes it lower case.
}
\examples{
clean_text("HELLO_World!!!")
}
\seealso{
\link{best_match}
}

