% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_scalar_in.R
\name{is.scalar_in}
\alias{is.scalar_in}
\alias{is.scalar_in01}
\title{Test if scalar is in intervall}
\usage{
is.scalar_in(left, right)

is.scalar_in01(x)
}
\arguments{
\item{left, right}{arguments passed to \code{\link[dplyr]{between}}}

\item{x}{R object to be tested, most likely a numeric vector of length one 
(other formats are allowed but will always return \code{FALSE}).}
}
\value{
\code{is.scalar_in01} returns \code{TRUE} if \code{x} is an atomic vector of 
length one and \code{0 <= as_numeric(x) <= 1}.
\code{is.scalar_in} return a function similair to \code{is.scalar_in01} but with 
specified boundaries.
}
\description{
Test if scalar is in intervall
}
\examples{
is.scalar_in01(.5) # TRUE
is.scalar_in01(5) # FALSE
is.scalar_in01(seq(0,1,.1)) # FALSE

is_scalar_in09 <- is.scalar_in(0,9)
is_scalar_in09(5) # TRUE
}

