% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lownames.R
\name{lownames}
\alias{lownames}
\title{Make all names in data.frame lower case}
\usage{
lownames(df)
}
\arguments{
\item{df}{A data.frame, possibly with some names with capital letters}
}
\value{
\code{df} is returned unchanged, except that capital letters in names are changed to lower case.
}
\description{
Tests are also performed so that all column names will stay unique!
}
\examples{
df <- data.frame(Hello = 1:10, World = 1:10)
lownames(df)
}

