% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_r_script.R
\name{make_r_script}
\alias{make_r_script}
\title{Dump skript together with all functions from package}
\usage{
make_r_script(script = NULL, package, outfile = "./inca_r_script.R",
  all = TRUE)
}
\arguments{
\item{script}{connection with script (file) to append tp function defenitions}

\item{package}{name of package (as character)}

\item{outfile}{filename for dump file}

\item{all}{should all (even non exported) objects from the package be exported?}
}
\value{
nothing (function called for its side effects)
}
\description{
Dump skript together with all functions from package
}

