% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psum.R
\name{psum}
\alias{psum}
\title{Parallel sum}
\usage{
psum(..., na.rm = FALSE)
}
\arguments{
\item{...}{numeric vectors}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\description{
This functin is to \link{sum}, what \link{pmin} and \link{pmax} is to \link{min} and \link{max}.
}
\examples{
psum(1:10, 1:10, 1:10) 
}

