% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rccmisc.R
\docType{package}
\name{rccmisc}
\alias{rccmisc}
\alias{rccmisc-package}
\title{Miscellaneous R functions for the Regional Cancer Centers}
\description{
The content of the package is quite diversed. It contains various functions used for various purposes.
}
\details{
Use \code{help(package="rccmisc")} to list all (exported) functions from the package.
}
\author{
Erik Bulow
}

