% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/width.R
\name{width}
\alias{width}
\title{Calculate the width of the range of x}
\usage{
width(x)
}
\arguments{
\item{x}{object to calculate range for}
}
\value{
The width of the range of \code{x} as integer.
}
\description{
Calculate the width of the range of x
}
\examples{
width(1:10)
width(c(6748, 234, 2456, 5678))
width(sample(345))
}

